refactor: move TIdent into separate files
