+++ src/meta/app/src/background/job_ident.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::tenant_key::TIdent;
+use crate::tenant_key::ident::TIdent;
 
 /// Defines the meta-service key for background job.
 pub type BackgroundJobIdent = TIdent<Resource>;
@@ -25,7 +25,7 @@ mod kvapi_impl {
     use databend_common_meta_kvapi::kvapi::Key;
 
     use crate::background::BackgroundJobId;
-    use crate::tenant_key::TenantResource;
+    use crate::tenant_key::resource::TenantResource;
 
     pub struct Resource;
     impl TenantResource for Resource {
