+++ src/meta/app/src/principal/user_stage_ident.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::tenant_key::TIdent;
+use crate::tenant_key::ident::TIdent;
 
 /// Define the meta-service key for a stage.
 pub type StageIdent = TIdent<Resource>;
@@ -24,7 +24,7 @@ mod kvapi_impl {
     use databend_common_meta_kvapi::kvapi;
 
     use crate::principal::StageInfo;
-    use crate::tenant_key::TenantResource;
+    use crate::tenant_key::resource::TenantResource;
 
     pub struct Resource;
     impl TenantResource for Resource {
